+++ src/main/java/gift/product/category/dto/request/UpdateCategoryRequest.java
@@ -1,11 +1,11 @@
-package gift.product.dto.category.request;
+package gift.product.category.dto.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
-public record CreateCategoryRequest(
-    @NotNull
+public record UpdateCategoryRequest(
     @NotEmpty
+    @NotNull
     String name
 ) {
