+++ src/main/java/gift/product/category/dto/response/CategoryResponse.java
@@ -1,4 +1,4 @@
-package gift.product.dto.category.response;
+package gift.product.category.dto.response;
 
 public record CategoryResponse(
     Long id,
