+++ src/main/java/gift/product/category/service/CategoryService.java
@@ -1,12 +1,12 @@
-package gift.product.service;
+package gift.product.category.service;
 
 import gift.exception.category.CategoryAlreadyExistException;
 import gift.exception.category.CategoryNotFoundException;
-import gift.product.dto.category.request.CreateCategoryRequest;
-import gift.product.dto.category.request.UpdateCategoryRequest;
-import gift.product.dto.category.response.CategoryResponse;
-import gift.product.entity.Category;
-import gift.product.repository.CategoryRepository;
+import gift.product.category.dto.request.CreateCategoryRequest;
+import gift.product.category.dto.request.UpdateCategoryRequest;
+import gift.product.category.dto.response.CategoryResponse;
+import gift.product.category.entity.Category;
+import gift.product.category.repository.CategoryRepository;
 import gift.util.mapper.CategoryMapper;
 import java.util.List;
 import org.springframework.stereotype.Service;
