+++ src/main/java/gift/product/service/ProductService.java
@@ -5,9 +5,9 @@ import gift.exception.product.ProductNotFoundException;
 import gift.product.dto.request.CreateProductRequest;
 import gift.product.dto.request.UpdateProductRequest;
 import gift.product.dto.response.ProductResponse;
-import gift.product.entity.Category;
+import gift.product.category.entity.Category;
 import gift.product.entity.Product;
-import gift.product.repository.CategoryRepository;
+import gift.product.category.repository.CategoryRepository;
 import gift.product.repository.ProductRepository;
 import gift.util.mapper.ProductMapper;
 import org.springframework.data.domain.Page;
