+++ src/test/java/gift/unit/service/CategoryServiceTest.java
@@ -9,12 +9,12 @@ import static org.mockito.BDDMockito.willDoNothing;
 import static org.mockito.Mockito.times;
 
 import gift.exception.category.CategoryNotFoundException;
-import gift.product.dto.category.request.CreateCategoryRequest;
-import gift.product.dto.category.request.UpdateCategoryRequest;
-import gift.product.dto.category.response.CategoryResponse;
-import gift.product.entity.Category;
-import gift.product.repository.CategoryRepository;
-import gift.product.service.CategoryService;
+import gift.product.category.dto.request.CreateCategoryRequest;
+import gift.product.category.dto.request.UpdateCategoryRequest;
+import gift.product.category.dto.response.CategoryResponse;
+import gift.product.category.entity.Category;
+import gift.product.category.repository.CategoryRepository;
+import gift.product.category.service.CategoryService;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
