+++ src/test/java/gift/unit/service/ProductServiceTest.java
@@ -12,9 +12,9 @@ import gift.exception.product.ProductNotFoundException;
 import gift.product.dto.request.CreateProductRequest;
 import gift.product.dto.request.UpdateProductRequest;
 import gift.product.dto.response.ProductResponse;
-import gift.product.entity.Category;
+import gift.product.category.entity.Category;
 import gift.product.entity.Product;
-import gift.product.repository.CategoryRepository;
+import gift.product.category.repository.CategoryRepository;
 import gift.product.repository.ProductRepository;
 import gift.product.service.ProductService;
 import gift.util.mapper.ProductMapper;
